fix: prevent crash when syncing library podcasts
add YU shield icon for yutorah provider

fix: raise SetupFailedError on login failure

refactor: split into multiple files, require login

fix: remove LIBRARY_PODCASTS and get_library_podcasts override

small fixes

fix: split some files and pr review comments

small logical fixes
